refactor notes components
