+++ enjoy/src/renderer/components/index.ts
@@ -2,15 +2,13 @@ export * from "./audios";
 export * from "./conversations";
 export * from "./meanings";
 export * from "./messages";
+export * from "./medias";
+export * from "./notes";
+export * from "./posts";
 export * from "./preferences";
 export * from "./pronunciation-assessments";
 export * from "./recordings";
 export * from "./stories";
-export * from "./videos";
-
-export * from "./medias";
-
-export * from "./posts";
 export * from "./users";
-
+export * from "./videos";
 export * from "./widgets";
