+++ enjoy/src/renderer/components/notes/index.ts
@@ -1,0 +1,2 @@
+export * from './note-card';
+export * from './note-form';
